First attempt at a simple GUI

+++ src/data/SPLStandardMessage.java
@@ -12,8 +12,8 @@ public class SPLStandardMessage implements Serializable {
      * Some constants from the C-structure.
      */
     public static final String SPL_STANDARD_MESSAGE_STRUCT_HEADER = "SPL ";
-    public static final int SPL_STANDARD_MESSAGE_STRUCT_VERSION = 5;
-    public static final int SPL_STANDARD_MESSAGE_DATA_SIZE = 800;
+    public static final byte SPL_STANDARD_MESSAGE_STRUCT_VERSION = 5;
+    public static final short SPL_STANDARD_MESSAGE_DATA_SIZE = 800;
     public static final int SIZE = 4 // header size
                                    + 1 // byte for the version
                                    + 1 // player number
